+++ physicsnemo/datapipes/cae/__init__.py
@@ -16,3 +16,4 @@
 
 from .domino_datapipe import DoMINODataPipe
 from .mesh_datapipe import MeshDatapipe
+from .transient_mesh_datapipe import TransientMeshDatapipe